Rename research_plans folder and all subfolders to camelCase
