+++ app/packs/src/components/elements/researchPlans/ResearchPlanDetails.js
@@ -9,17 +9,17 @@ import ElementActions from 'src/stores/alt/actions/ElementActions';
 import DetailActions from 'src/stores/alt/actions/DetailActions';
 import ResearchPlansFetcher from 'src/fetchers/ResearchPlansFetcher';
 import ResearchPlansLiteratures from 'src/components/elements/DetailsTabLiteratures';
-import ResearchPlanWellplates from 'src/components/elements/research_plans/wellplates_tab/ResearchPlanWellplates';
-import ResearchPlanMetadata from 'src/components/elements/research_plans/ResearchPlanMetadata';
+import ResearchPlanWellplates from 'src/components/elements/researchPlans/wellplatesTab/ResearchPlanWellplates';
+import ResearchPlanMetadata from 'src/components/elements/researchPlans/ResearchPlanMetadata';
 import Attachment from 'src/models/Attachment';
 import Utils from 'src/utilities/Functions';
 import LoadingActions from 'src/stores/alt/actions/LoadingActions';
 import ConfirmClose from 'src/components/common/ConfirmClose';
 import ResearchPlan from 'src/models/ResearchPlan';
-import ResearchPlanDetailsAttachments from 'src/components/elements/research_plans/attachments_tab/ResearchPlanDetailsAttachments';
-import ResearchPlanDetailsBody from 'src/components/elements/research_plans/research_plan_tab/ResearchPlanDetailsBody';
-import ResearchPlanDetailsName from 'src/components/elements/research_plans/research_plan_tab/ResearchPlanDetailsName';
-import ResearchPlanDetailsContainers from 'src/components/elements/research_plans/analyses_tab/ResearchPlanDetailsContainers';
+import ResearchPlanDetailsAttachments from 'src/components/elements/researchPlans/attachmentsTab/ResearchPlanDetailsAttachments';
+import ResearchPlanDetailsBody from 'src/components/elements/researchPlans/researchPlanTab/ResearchPlanDetailsBody';
+import ResearchPlanDetailsName from 'src/components/elements/researchPlans/researchPlanTab/ResearchPlanDetailsName';
+import ResearchPlanDetailsContainers from 'src/components/elements/researchPlans/analysesTab/ResearchPlanDetailsContainers';
 import ElementDetailSortTab from 'src/components/elements/ElementDetailSortTab';
 import { addSegmentTabs } from 'src/components/generic/SegmentDetails';
 import PrivateNoteElement from 'src/components/elements/PrivateNoteElement';
