+++ app/packs/src/components/elements/researchPlans/analysesTab/ResearchPlanDetailsContainers.js
