+++ app/packs/src/components/elements/researchPlans/researchPlanTab/ResearchPlanDetailsDragSource.js
