+++ app/packs/src/components/elements/researchPlans/researchPlanTab/ResearchPlanDetailsDropTarget.js
