+++ app/packs/src/components/elements/researchPlans/researchPlanTab/ResearchPlanDetailsField.js
@@ -2,14 +2,14 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { Button, ButtonGroup, Row, Col, ControlLabel, Tooltip, OverlayTrigger, DropdownButton, MenuItem } from 'react-bootstrap';
-import ResearchPlanDetailsDragSource from 'src/components/elements/research_plans/research_plan_tab/ResearchPlanDetailsDragSource';
-import ResearchPlanDetailsDropTarget from 'src/components/elements/research_plans/research_plan_tab/ResearchPlanDetailsDropTarget';
-import ResearchPlanDetailsFieldRichText from 'src/components/elements/research_plans/research_plan_tab/ResearchPlanDetailsFieldRichText';
-import ResearchPlanDetailsFieldKetcher from 'src/components/elements/research_plans/research_plan_tab/ResearchPlanDetailsFieldKetcher';
-import ResearchPlanDetailsFieldImage from 'src/components/elements/research_plans/research_plan_tab/ResearchPlanDetailsFieldImage';
-import ResearchPlanDetailsFieldTable from 'src/components/elements/research_plans/research_plan_tab/ResearchPlanDetailsFieldTable';
-import ResearchPlanDetailsFieldSample from 'src/components/elements/research_plans/research_plan_tab/ResearchPlanDetailsFieldSample';
-import ResearchPlanDetailsFieldReaction from 'src/components/elements/research_plans/research_plan_tab/ResearchPlanDetailsFieldReaction';
+import ResearchPlanDetailsDragSource from 'src/components/elements/researchPlans/researchPlanTab/ResearchPlanDetailsDragSource';
+import ResearchPlanDetailsDropTarget from 'src/components/elements/researchPlans/researchPlanTab/ResearchPlanDetailsDropTarget';
+import ResearchPlanDetailsFieldRichText from 'src/components/elements/researchPlans/researchPlanTab/ResearchPlanDetailsFieldRichText';
+import ResearchPlanDetailsFieldKetcher from 'src/components/elements/researchPlans/researchPlanTab/ResearchPlanDetailsFieldKetcher';
+import ResearchPlanDetailsFieldImage from 'src/components/elements/researchPlans/researchPlanTab/ResearchPlanDetailsFieldImage';
+import ResearchPlanDetailsFieldTable from 'src/components/elements/researchPlans/researchPlanTab/ResearchPlanDetailsFieldTable';
+import ResearchPlanDetailsFieldSample from 'src/components/elements/researchPlans/researchPlanTab/ResearchPlanDetailsFieldSample';
+import ResearchPlanDetailsFieldReaction from 'src/components/elements/researchPlans/researchPlanTab/ResearchPlanDetailsFieldReaction';
 
 export default class ResearchPlanDetailsField extends Component {
   render() {
