+++ app/packs/src/components/elements/researchPlans/researchPlanTab/ResearchPlanDetailsFieldImage.js
