+++ app/packs/src/components/elements/researchPlans/researchPlanTab/ResearchPlanDetailsFieldReaction.js
