+++ app/packs/src/components/elements/researchPlans/researchPlanTab/ResearchPlanDetailsFieldRichText.js
