+++ app/packs/src/components/elements/researchPlans/researchPlanTab/ResearchPlanDetailsFieldSample.js
