+++ app/packs/src/components/elements/researchPlans/researchPlanTab/ResearchPlanDetailsName.js
